+++ apps/native-component-list/src/screens/UIUniversal/UIUniversalScreen.tsx
@@ -80,11 +80,35 @@ export const UIUniversalScreens = [
     },
   },
   {
+    name: `${SCREEN_NAME_PREFIX}Collapsible`,
+    route: 'ui-universal/collapsible',
+    options: { title: 'Collapsible' },
+    getComponent() {
+      return optionalRequire(() => require('./CollapsibleScreen'));
+    },
+  },
+  {
     name: `${SCREEN_NAME_PREFIX}FieldGroup`,
     route: 'ui-universal/field-group',
     options: { title: 'FieldGroup' },
     getComponent() {
       return optionalRequire(() => require('./FieldGroupScreen'));
+    },
+  },
+  {
+    name: `${SCREEN_NAME_PREFIX}List`,
+    route: 'ui-universal/list',
+    options: { title: 'List' },
+    getComponent() {
+      return optionalRequire(() => require('./ListScreen'));
+    },
+  },
+  {
+    name: `${SCREEN_NAME_PREFIX}Picker`,
+    route: 'ui-universal/picker',
+    options: { title: 'Picker' },
+    getComponent() {
+      return optionalRequire(() => require('./PickerScreen'));
     },
   },
   {